Update README.md
I loved the series back then before it ended. What made me love the show was the lore and background of the characters and the world they are in. From the start the shows story and premise is simple, but once you get further in the series you get to see the plot uncover. Revealing a more interesting lore and story of what is going on.